Implements loadGraph(), addNode(), and addEdge()
Implements these methods by reading a file and breaking the input up into an array that correctly distributes them to their intended type. Uses file information to add nodes to the nodes array and implements a hash map to map each array index(key) to the name of the city it represents. The values are then used to create and adjacency list (array of linked lists, or nodes with next pointers) to show which cities connect, and the cost of traveling from both the origin city as well as the destination city.