Implements Dijkstra's table, Dijkstra, and PriorityQueue
I forgot to commit after completing the table and PriorityQueue. This commit has a lot of changes for that reason. I implemented the priority queue using the code example from github. I added some methods and made some changes to exisiting ones in order to make it more appropriate for this project. methods I added include isEmpty() and bubbleUp() which helps with the reduceKey() method. I also made a public getter method for the adjacency list in the graph class in order to make it easier to find edges and costs for Dijkstra. This has all been thoroughly tested. In this commit I have implemented my first attempt at computing the shortest path by backtracking but it has not yet been tested.